+++ packages/apollo-link-batch-http/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### vNext
 
+- Add `createBatchLink` option to customise link creation.
+
 ### 1.2.2
 - Update apollo-link [#559](https://github.com/apollographql/apollo-link/pull/559)
 - Check for signal already present on `fetchOptions` [#584](https://github.com/apollographql/apollo-link/pull/584)